+++ infra/modules/providers/azure/ad-application/main.tf
@@ -12,22 +12,20 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
+resource "random_uuid" "app_role_id" {} // https://registry.terraform.io/providers/hashicorp/azuread/latest/docs/guides/microsoft-graph#new-required-uuid-fields
+
 data "azuread_service_principal" "main" {
   count        = var.aad_client_id != "" ? 0 : length(local.api_names)
   display_name = local.api_names[count.index]
 }
 
 resource "azuread_application" "main" {
-  count                      = var.aad_client_id != "" ? 0 : 1
-  name                       = var.name
-  homepage                   = coalesce(var.homepage, local.homepage)
-  identifier_uris            = local.identifier_uris
-  reply_urls                 = var.reply_urls
-  available_to_other_tenants = var.available_to_other_tenants
-  public_client              = local.public_client
-  oauth2_allow_implicit_flow = var.oauth2_allow_implicit_flow
-  group_membership_claims    = var.group_membership_claims
-  type                       = local.type
+  count                          = var.aad_client_id != "" ? 0 : 1
+  display_name                   = var.display_name
+  identifier_uris                = local.identifier_uris
+  sign_in_audience               = var.sign_in_audience
+  fallback_public_client_enabled = local.fallback_public_client_enabled
+  group_membership_claims        = var.group_membership_claims
 
   dynamic "required_resource_access" {
     for_each = local.required_resource_access
@@ -51,10 +49,42 @@ resource "azuread_application" "main" {
 
     content {
       allowed_member_types = app_role.value.member_types
-      display_name         = app_role.value.name
+      display_name         = app_role.value.display_name
       description          = app_role.value.description
       value                = coalesce(app_role.value.value, app_role.value.name)
-      is_enabled           = app_role.value.enabled
+      enabled              = app_role.value.enabled
+      id                   = random_uuid.app_role_id.result
+    }
+  }
+
+  api {
+    mapped_claims_enabled          = false
+    requested_access_token_version = 1
+
+    known_client_applications = var.known_client_applications
+
+    dynamic "oauth2_permission_scope" {
+      for_each = local.oauth2_permission_scopes
+      content {
+        admin_consent_description  = oauth2_permission_scope.value.admin_consent_description
+        admin_consent_display_name = oauth2_permission_scope.value.admin_consent_display_name
+        enabled                    = oauth2_permission_scope.value.enabled
+        id                         = oauth2_permission_scope.value.id
+        type                       = oauth2_permission_scope.value.type
+        user_consent_description   = oauth2_permission_scope.value.user_consent_description
+        user_consent_display_name  = oauth2_permission_scope.value.user_consent_display_name
+        value                      = oauth2_permission_scope.value.value
+      }
+    }
+  }
+
+  web {
+    homepage_url  = coalesce(var.homepage_url, local.homepage_url)
+    redirect_uris = var.redirect_uris
+
+    implicit_grant {
+      access_token_issuance_enabled = var.access_token_issuance_enabled
+      id_token_issuance_enabled     = true
     }
   }
 }
@@ -69,7 +99,6 @@ resource "azuread_application_password" "main" {
   count                 = var.aad_client_id == "" && var.password != null ? 1 : 0
   application_object_id = var.aad_client_id != "" ? null : azuread_application.main[0].object_id
 
-  value             = coalesce(var.password, random_password.main[0].result)
   end_date          = local.end_date
   end_date_relative = local.end_date_relative