+++ infra/modules/providers/azure/service-principal/main.tf
@@ -25,9 +25,9 @@ data "azuread_service_principal" "main" {
 
 
 resource "azuread_application" "main" {
-  count                      = local.create_count
-  name                       = var.name
-  available_to_other_tenants = false
+  count            = local.create_count
+  display_name     = var.display_name
+  sign_in_audience = "AzureADMyOrg"
 
   dynamic "required_resource_access" {
     for_each = local.required_resource_access
@@ -63,7 +63,6 @@ resource "azuread_service_principal_password" "main" {
   count                = local.create_count != 0 && var.password != null ? 1 : 0
   service_principal_id = azuread_service_principal.main[0].id
 
-  value             = coalesce(var.password, random_password.main[0].result)
   end_date          = local.end_date
   end_date_relative = local.end_date_relative