+++ infra/templates/osdu-r3-mvp/service_resources/main.tf
@@ -34,7 +34,7 @@ terraform {
     }
     azuread = {
       source  = "hashicorp/azuread"
-      version = "=1.1.1"
+      version = "=2.33.0"
     }
     random = {
       source  = "hashicorp/random"